More const correctness. Can't have too much, you know.
Changed a long int endian converstion test to regular int because the
rather longer long int on flash we're causing the conversion operation
to come out differently and the test to fail.

+++ src/ds++/test/tstPacking_Utils.cc
@@ -501,9 +501,9 @@ void endian_conversion_test()
     Unpacker up(true);
 
 
-    // Test the long int type.
-    const long int moo = 0xDEADBEEF;
-    const int length = sizeof(long int);
+    // Test the int type.
+    const int moo = 0xDEADBEEF;
+    const int length = sizeof(int);
 
     // Pack
     char data[length];
@@ -511,7 +511,7 @@ void endian_conversion_test()
     p << moo;
 
     // Unpack
-    long int oom = 0; 
+    int oom = 0; 
     up.set_buffer(length, data);
     up >> oom;
